commit to change branch

+++ Application/h_bridge_task/h_bridge_sequence.h
@@ -20,12 +20,12 @@ typedef struct _H_Bridge_task_typedef_
     uint8_t     hv_pos_count;
     uint8_t     hv_neg_count;
 
-    uint8_t     hv_delay_ms;
+    uint16_t    hv_delay_ms;
 
-    uint8_t     hv_pos_on_ms;
-    uint8_t     hv_pos_off_ms;
-    uint8_t     hv_neg_on_ms;
-    uint8_t     hv_neg_off_ms;
+    uint16_t    hv_pos_on_ms;
+    uint16_t    hv_pos_off_ms;
+    uint16_t    hv_neg_on_ms;
+    uint16_t    hv_neg_off_ms;
 
     uint8_t     lv_pos_count;
     uint8_t     lv_neg_count;
